fix: apply circuit breaker before Http().singleRequest, not after
Move CircuitBreaker creation from ElasticSearchResponseHandler to
ElasticSearchClient so it wraps Http().singleRequest() directly inside
withConcurrencyLimit. Previously the breaker wrapped an already-started
Future, so it could not prevent requests from being sent to ES when open.
Also distinguish CircuitBreakerOpenException in error log (warn vs error).
